requirements/base.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
Django==6.0
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-spectacular==0.29.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Markdown==3.10
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
pip-review==1.3.0
psycopg2==2.9.11
pycparser==2.23
PyJWT==2.10.1
python3-openid==3.2.0
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
setuptools==80.9.0
smmap==5.0.2
social-auth-app-django==5.7.0
social-auth-core==4.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
